Moved graphics-related to own directory.
