+++ VoxelEngine/VoxelEngine/VoxelEngine/Graphics/GraphicsIncludes.h
@@ -1,0 +1,13 @@
+#include <GL/glew.h>
+#include <GLFW/glfw3.h>
+#include <MathIncludes.h>
+
+#include <Graphics/abo.h>
+#include <Graphics/dib.h>
+#include <Graphics/ibo.h>
+#include <Graphics/utilities.h>
+#include <Graphics/shader.h>
+#include <Graphics/vao.h>
+#include <Graphics/vbo.h>
+#include <Graphics/param_bo.h>
+#include <Graphics/Vertices.h>
