+++ VoxelEngine/VoxelEngine/VoxelEngine/Graphics/abo.cpp
@@ -1,4 +1,4 @@
-#include <Systems/Graphics/GraphicsIncludes.h>
+#include <Graphics/GraphicsIncludes.h>
 
 ABO::ABO(GLuint num) : numCounters_(num)
 {
