+++ VoxelEngine/VoxelEngine/VoxelEngine/Graphics/dib.cpp
@@ -1,4 +1,4 @@
-#include <Systems/Graphics/GraphicsIncludes.h>
+#include <Graphics/GraphicsIncludes.h>
 
 DIB::DIB(void* data, GLsizei size, GLenum drawmode)
 {
