+++ VoxelEngine/VoxelEngine/VoxelEngine/Graphics/misc_utils.h
