+++ VoxelEngine/VoxelEngine/VoxelEngine/Graphics/utilities.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#include <Graphics/math_utils.h>
+#include <Graphics/misc_utils.h>
