+++ VoxelEngine/VoxelEngine/VoxelEngine/Graphics/vao.cpp
@@ -1,4 +1,4 @@
-#include <Systems/Graphics/GraphicsIncludes.h>
+#include <Graphics/GraphicsIncludes.h>
 
 VAO::VAO()
 {
