+++ VoxelEngine/VoxelEngine/VoxelEngine/Graphics/vbo.cpp
@@ -1,4 +1,4 @@
-#include <Systems/Graphics/GraphicsIncludes.h>
+#include <Graphics/GraphicsIncludes.h>
 
 VBO::VBO(const void * data, unsigned int size, GLenum drawmode)
 {
