--- VoxelEngine/VoxelEngine/VoxelEngine/Headers/Systems/Graphics/GraphicsIncludes.h
@@ -1,13 +1,0 @@
-#include <GL/glew.h>
-#include <GLFW/glfw3.h>
-#include <MathIncludes.h>
-
-#include <Systems/Graphics/abo.h>
-#include <Systems/Graphics/dib.h>
-#include <Systems/Graphics/ibo.h>
-#include <Systems/Graphics/utilities.h>
-#include <Systems/Graphics/shader.h>
-#include <Systems/Graphics/vao.h>
-#include <Systems/Graphics/vbo.h>
-#include <Systems/Graphics/param_bo.h>
-#include <Systems/Graphics/Vertices.h>
