--- VoxelEngine/VoxelEngine/VoxelEngine/Headers/Systems/Graphics/utilities.h
@@ -1,4 +1,0 @@
-#pragma once
-
-#include "math_utils.h"
-#include "misc_utils.h"
