+++ VoxelEngine/VoxelEngine/VoxelEngine/VoxelEngine.vcxproj
@@ -29,6 +29,22 @@
     <ClInclude Include="3rdParty\imgui\imstb_rectpack.h" />
     <ClInclude Include="3rdParty\imgui\imstb_textedit.h" />
     <ClInclude Include="3rdParty\imgui\imstb_truetype.h" />
+    <ClInclude Include="Graphics\abo.h" />
+    <ClInclude Include="Graphics\BufferAllocator.h" />
+    <ClInclude Include="Graphics\Context.h" />
+    <ClInclude Include="Graphics\dib.h" />
+    <ClInclude Include="Graphics\GraphicsIncludes.h" />
+    <ClInclude Include="Graphics\ibo.h" />
+    <ClInclude Include="Graphics\math_utils.h" />
+    <ClInclude Include="Graphics\misc_utils.h" />
+    <ClInclude Include="Graphics\param_bo.h" />
+    <ClInclude Include="Graphics\shader.h" />
+    <ClInclude Include="Graphics\Shapes.h" />
+    <ClInclude Include="Graphics\utilities.h" />
+    <ClInclude Include="Graphics\vao.h" />
+    <ClInclude Include="Graphics\vbo.h" />
+    <ClInclude Include="Graphics\vbo_layout.h" />
+    <ClInclude Include="Graphics\Vertices.h" />
     <ClInclude Include="Headers\Components\AllComponentHeaders.h" />
     <ClInclude Include="Headers\Components\Camera.h" />
     <ClInclude Include="Headers\Components\Component.h" />
@@ -63,25 +79,9 @@
     <ClInclude Include="Headers\Systems\AllSystemHeaders.h" />
     <ClInclude Include="Headers\Systems\FrameRateController.h" />
     <ClInclude Include="Headers\Systems\GraphicsSystem.h" />
-    <ClInclude Include="Headers\Systems\Graphics\abo.h" />
-    <ClInclude Include="Headers\Systems\Graphics\BufferAllocator.h" />
-    <ClInclude Include="Headers\Systems\Graphics\Context.h" />
-    <ClInclude Include="Headers\Systems\Graphics\dib.h" />
-    <ClInclude Include="Headers\Systems\Graphics\GraphicsIncludes.h" />
-    <ClInclude Include="Headers\Systems\Graphics\ibo.h" />
-    <ClInclude Include="Headers\Systems\Graphics\math_utils.h" />
-    <ClInclude Include="Headers\Systems\Graphics\misc_utils.h" />
-    <ClInclude Include="Headers\Systems\Graphics\param_bo.h" />
-    <ClInclude Include="Headers\Systems\Graphics\shader.h" />
-    <ClInclude Include="Headers\Systems\Graphics\Shapes.h" />
-    <ClInclude Include="Headers\Systems\Graphics\vao.h" />
-    <ClInclude Include="Headers\Systems\Graphics\vbo.h" />
-    <ClInclude Include="Headers\Systems\Graphics\vbo_layout.h" />
-    <ClInclude Include="Headers\Systems\Graphics\Vertices.h" />
     <ClInclude Include="Headers\Systems\InputSystem.h" />
     <ClInclude Include="Headers\Systems\StubSystem.h" />
     <ClInclude Include="Headers\Systems\System.h" />
-    <ClInclude Include="Source\Systems\Graphics\vbo_layout.h" />
     <ClInclude Include="Headers\Systems\TraceSystem.h" />
     <ClInclude Include="Voxels\biome.h" />
     <ClInclude Include="Voxels\block.h" />
@@ -118,7 +118,6 @@
     <ClInclude Include="Voxels\Utilities\Palette.h" />
     <ClInclude Include="Voxels\Utilities\serialize.h" />
     <ClInclude Include="Voxels\Utilities\Timer.h" />
-    <ClInclude Include="Voxels\Utilities\utilities.h" />
     <ClInclude Include="Voxels\WorldGen\WorldGen2.h" />
     <ClInclude Include="Voxels\World\chunk_manager.h" />
     <ClInclude Include="Voxels\World\chunk_manager_base.h" />
@@ -141,6 +140,16 @@
       <PreprocessorDefinitions Condition="'$(Configuration)|$(Platform)'=='Release|x64'">_CRT_SECURE_NO_WARNINGS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
     </ClCompile>
     <ClCompile Include="3rdParty\imgui\imgui_widgets.cpp" />
+    <ClCompile Include="Graphics\abo.cpp" />
+    <ClCompile Include="Graphics\Context.cpp" />
+    <ClCompile Include="Graphics\dib.cpp" />
+    <ClCompile Include="Graphics\ibo.cpp" />
+    <ClCompile Include="Graphics\math_utils.cpp" />
+    <ClCompile Include="Graphics\misc_utils.cpp" />
+    <ClCompile Include="Graphics\param_bo.cpp" />
+    <ClCompile Include="Graphics\shader.cpp" />
+    <ClCompile Include="Graphics\vao.cpp" />
+    <ClCompile Include="Graphics\vbo.cpp" />
     <ClCompile Include="Source\Components\Camera.cpp" />
     <ClCompile Include="Source\Components\Component.cpp" />
     <ClCompile Include="Source\Components\StubComponent.cpp" />
@@ -156,19 +165,6 @@
     <ClCompile Include="Source\GuiIfy.cpp" />
     <ClCompile Include="Source\Systems\FrameRateController.cpp" />
     <ClCompile Include="Source\Systems\GraphicsSystem.cpp" />
-    <ClCompile Include="Source\Systems\Graphics\abo.cpp" />
-    <ClCompile Include="Source\Systems\Graphics\Context.cpp" />
-    <ClCompile Include="Source\Systems\Graphics\dib.cpp" />
-    <ClCompile Include="Source\Systems\Graphics\ibo.cpp" />
-    <ClCompile Include="Source\Systems\Graphics\math_utils.cpp" />
-    <ClCompile Include="Source\Systems\Graphics\misc_utils.cpp" />
-    <ClCompile Include="Source\Systems\Graphics\param_bo.cpp" />
-    <ClCompile Include="Source\Systems\Graphics\shader.cpp">
-      <PreprocessorDefinitions Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">_CRT_SECURE_NO_WARNINGS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
-      <PreprocessorDefinitions Condition="'$(Configuration)|$(Platform)'=='Release|x64'">_CRT_SECURE_NO_WARNINGS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
-    </ClCompile>
-    <ClCompile Include="Source\Systems\Graphics\vao.cpp" />
-    <ClCompile Include="Source\Systems\Graphics\vbo.cpp" />
     <ClCompile Include="Source\Systems\InputSystem.cpp" />
     <ClCompile Include="Source\Systems\StubSystem.cpp" />
     <ClCompile Include="Source\Systems\TraceSystem.cpp" />
@@ -202,7 +198,7 @@
     <ClCompile Include="Voxels\World\World.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="Headers\Systems\Graphics\BufferAllocator.inl" />
+    <None Include="Graphics\BufferAllocator.inl" />
     <None Include="Voxels\Chunks\BlockStorage.inl" />
     <None Include="Voxels\Chunks\ChunkHelpers.inl" />
     <None Include="Voxels\Utilities\BufferAllocator.inl" />
@@ -270,7 +266,7 @@
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <LinkIncremental>true</LinkIncremental>
-    <IncludePath>$(ProjectDir)3rdParty;$(ProjectDir)Headers;$(ProjectDir)3rdParty\glew-2.1.0\include;$(ProjectDir)3rdParty\glfw-3.3-win64\include;$(VC_IncludePath);$(WindowsSDK_IncludePath)</IncludePath>
+    <IncludePath>$(ProjectDir)3rdParty;$(ProjectDir)Headers;$(ProjectDir)3rdParty\glew-2.1.0\include;$(ProjectDir);$(ProjectDir)3rdParty\glfw-3.3-win64\include;$(VC_IncludePath);$(WindowsSDK_IncludePath)</IncludePath>
     <PreBuildEventUseInBuild>true</PreBuildEventUseInBuild>
     <CustomBuildBeforeTargets>
     </CustomBuildBeforeTargets>
@@ -287,7 +283,7 @@
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <LinkIncremental>false</LinkIncremental>
-    <IncludePath>$(ProjectDir)3rdParty;$(ProjectDir)Headers;$(ProjectDir)3rdParty\glew-2.1.0\include;$(ProjectDir)3rdParty\glfw-3.3-win64\include;$(ProjectDir)Voxels;$(ProjectDir)3rdParty\cereal-1.2.2\include;$(ProjectDir)3rdParty\zlib-1.2.11;$(IncludePath)</IncludePath>
+    <IncludePath>$(ProjectDir)3rdParty;$(ProjectDir)Headers;$(ProjectDir)3rdParty\glew-2.1.0\include;$(ProjectDir);$(ProjectDir)3rdParty\glfw-3.3-win64\include;$(ProjectDir)Voxels;$(ProjectDir)3rdParty\cereal-1.2.2\include;$(ProjectDir)3rdParty\zlib-1.2.11;$(IncludePath)</IncludePath>
     <PreBuildEventUseInBuild>true</PreBuildEventUseInBuild>
     <CustomBuildBeforeTargets>
     </CustomBuildBeforeTargets>
