--- VoxelEngine/VoxelEngine/VoxelEngine/Voxels/Utilities/utilities.h
@@ -1,5 +1,0 @@
-// includes every utilities header
-#pragma once
-
-#include "math_utils.h"
-#include <Systems/Graphics/misc_utils.h>
