+++ VoxelEngine/VoxelEngine/VoxelEngine/Voxels/biome.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <Systems/Graphics/misc_utils.h>
+#include <Graphics/misc_utils.h>
 
 enum class TerrainType : unsigned
 {
